+++ examples/coloured/gradient.rs
@@ -1,20 +1,19 @@
 // [dependencies]
 // kdam = { version = "0.2.7", features = ["gradient", "template"] }
 
-use kdam::{tqdm, BarExt};
+use kdam::term::Colorizer;
+use kdam::tqdm;
 
 fn main() {
-    let mut pb = tqdm!(
+    for _ in tqdm!(
+        0..300,
         total = 300,
-        bar_format = "{animation} {percentage:3.0}%",
+        bar_format = format!("{{animation}} {}", "{percentage:3.0}%".colorize("#EE6FF8")),
         colour = "gradient(#5A56E0,#EE6FF8)",
         force_refresh = true
-    );
-
-    for _ in 0..300 {
+    ) {
         std::thread::sleep(std::time::Duration::from_secs_f32(0.02));
-        pb.update(1);
     }
 
-    eprint!("\n");
+    eprintln!();
 }