+++ src/styles/format.rs
@@ -26,14 +26,15 @@ pub fn format_sizeof(num: f64, divisor: f64) -> String {
 
 pub fn format_time(num: f64) -> String {
     let mut value = num;
-    let units = [(60., "s"), (60., "min"), (24., "hr")];
-    for (d, i) in units {
+
+    for (d, i) in [(60., "s"), (60., "min"), (24., "hr")] {
         if value.abs() < d - 0.005 {
-            return format!("{value:1.2}{i}");
+            return format!("{:1.2}{}", value, i);
         }
         value /= d;
     }
-    format!("{value:1.2}days")
+
+    format!("{:1.2}days", value)
 }
 
 /// Formats a number of seconds as a clock time, \[H:\]MM:SS and SSs.
@@ -60,4 +61,3 @@ pub fn format_interval(seconds: usize, human: bool) -> String {
 //     let n = format!("{}", n).to_string();
 //     return if f.len() < n.len() { f } else { n };
 // }
-